refactor: split ts plugin / ts program proxy logic
