+++ packages/vscode-vue-languageservice/src/index.ts
@@ -3,6 +3,10 @@ export { getTriggerCharacters } from './services/completion';
 export { margeWorkspaceEdits } from './services/rename';
 export { isRefType, isBlacklistNode } from './services/refAutoClose';
 export * from './utils/sourceMaps';
+export * from './documentService';
 export * from './languageService';
 export * from './sourceFile';
 export * from './types';
+export * from './basicRuntime';
+export * from './tsPluginProxy';
+export * from './tsProgramProxy';
